pyproject.toml
[tool.poetry]
name = "configer"
version = "0.0.1"
description = ""
authors = ["Armand Sauzay <sauzayarmand@gmail.com>"]
readme = "README.md"
packages = [{include = "configer"}]

[tool.poetry.dependencies]
python = ">=3.10,<4"
pydantic-settings = "^2.2.1"
scikit-learn = "^1.5.0"
pyyaml = "^6.0.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
